before refactoring
