+++ integrationtesting/main_test.go
@@ -1,4 +1,4 @@
-package integration
+package integrationtesting
 
 import (
 	"log"
